Removed many instances of strcpy, strcat and sprintf Added mulitple unit tests Corrected some unit tests to run on Windows Corrected the SnapScreen() action to work on scaled displays Added Query(dirs) and Query(specialcharacters) Got Query(windowid) to appear only on the right products Fixed XTWINOPS bug on Windows Corrected initial $PATH set-up on Windows and how x3270if is found by Prompt() Improved wc3270 Session Wizard behavior on Windows Terminal Added EXTRA_FLAGS to Windows pr3287 Makefile

+++ Common/b3270/winops.c
@@ -418,8 +418,8 @@ xtwinops(unsigned short p1, unsigned short *p2, unsigned short *p3, unsigned sho
     case XTW_3MOVE: /* Move to x,y */
 	ui_leaf(IndWindowChange,
 		AttrOperation, AT_STRING, WinMove,
-		AttrX, AT_INT, *p2,
-		AttrY, AT_INT, *p3,
+		AttrX, AT_INT, (uint64_t)*p2,
+		AttrY, AT_INT, (uint64_t)*p3,
 		NULL);
 	break;
     case XTW_4RESIZE_PIXELS: /* resize to window to h,w pixels */
@@ -427,20 +427,20 @@ xtwinops(unsigned short p1, unsigned short *p2, unsigned short *p3, unsigned sho
 	    ui_leaf(IndWindowChange,
 		    AttrOperation, AT_STRING, WinSize,
 		    AttrType, AT_STRING, SizeWindow,
-		    AttrHeight, AT_INT, *p2,
-		    AttrWidth, AT_INT, *p3,
+		    AttrHeight, AT_INT, (uint64_t)*p2,
+		    AttrWidth, AT_INT, (uint64_t)*p3,
 		    NULL);
 	} else if (p2) {
 	    ui_leaf(IndWindowChange,
 		    AttrOperation, AT_STRING, WinSize,
 		    AttrType, AT_STRING, SizeWindow,
-		    AttrHeight, AT_INT, *p2,
+		    AttrHeight, AT_INT, (uint64_t)*p2,
 		    NULL);
 	} else {
 	    ui_leaf(IndWindowChange,
 		    AttrOperation, AT_STRING, WinSize,
 		    AttrType, AT_STRING, SizeWindow,
-		    AttrWidth, AT_INT, *p3,
+		    AttrWidth, AT_INT, (uint64_t)*p3,
 		    NULL);
 	}
 	break;
